[Refactor] ViewModel 하나로 통합
