+++ RefillStation/RefillStation/Presentation/StoreDetailScene/ViewModel/StoreDetailViewModel.swift
@@ -8,74 +8,66 @@
 import UIKit
 
 final class StoreDetailViewModel {
-    let detailReviewViewModel: DetailReviewViewModel
-    let votedTagViewModel: VotedTagViewModel
-    let storeDetailInfoViewModel: StoreDetailInfoViewModel
-    let productListViewModel: ProductListViewModel
-    let operationInfoViewModel: OperationInfoViewModel
+    var mode: TabBarMode = .productLists
+    var store = MockEntityData.stores().first!
+    var detailReviews = MockEntityData.detailReviews()
+    var totalVoteCount = 5
+    var tagReviews = MockEntityData.tagReviews()
 
-    let storeDetailInfoViewHeight: CGFloat = 400
-    var mode: Mode = .productLists
-
-    var withoutReviewCount: Int {
-        if votedTagViewModel.totalVoteCount == 0 {
-            return 1
-        } else {
-            return 4
-        }
+    private let fetchProductListUseCase: FetchProductListUseCaseInterface
+    private(set) var categories = [ProductCategory]()
+    private(set) var currentCategoryFilter = ProductCategory.all
+    var filteredProducts: [Product] {
+        let filtered = products.filter({
+            if currentCategoryFilter == ProductCategory.all {
+                return true
+            } else {
+                return $0.category == currentCategoryFilter
+            }
+        })
+        return filtered
     }
 
-    var reloadItemsAt: (([IndexPath]) -> Void)? {
-        didSet {
-            productListViewModel.reloadItemsAt = { [weak self] indexPaths in
-                self?.reloadItemsAt?(indexPaths)
+    var operationInfos = MockEntityData.operations()
+    var products: [Product] = MockEntityData.products()
+
+    var operationInfoSeeMoreIndexPaths = Set<IndexPath>()
+    private var productListLoadTask: Cancellable?
+
+    init(fetchProductListUseCase: FetchProductListUseCaseInterface) {
+        self.fetchProductListUseCase = fetchProductListUseCase
+        products.forEach {
+            if !categories.contains($0.category) {
+                categories.append($0.category)
             }
         }
     }
 
-    func productListSection(for indexPath: IndexPath) -> ProductListSection {
-        if indexPath.row == 0 {
-            return ProductListSection.productCategory
-        } else if indexPath.row == 1 {
-            return ProductListSection.productsCount
-        } else {
-            return ProductListSection.productList
-        }
+    func categoryButtonDidTapped(category: ProductCategory?) {
+        guard let category = category else { return }
+        currentCategoryFilter = category
     }
 
-    func reviewListSection(for indexPath: IndexPath) -> ReviewSection {
-        if votedTagViewModel.totalVoteCount == 0 {
-            return ReviewSection.moveToWriteReview
-        } else if indexPath.row == 0 {
-            return ReviewSection.moveToWriteReview
-        } else if indexPath.row == 1 {
-            return ReviewSection.votedCount
-        } else if indexPath.row == 2 {
-            return ReviewSection.votedTag
-        } else if indexPath.row == 3 {
-            return ReviewSection.detailReviewCount
-        } else {
-            return ReviewSection.detailReviews
-        }
+    private func fetchProductList(storeId: Int, completion: @escaping (Result<[Product], Error>) -> Void) {
+        productListLoadTask = fetchProductListUseCase
+            .execute(requestValue: FetchProductListRequestValue(storeId: storeId)) { result in
+                switch result {
+                case .success(let products):
+                    completion(.success(products))
+                case .failure(let error):
+                    completion(.failure(error))
+                }
+            }
     }
 
-    init(
-        detailReviewViewModel: DetailReviewViewModel,
-        votedTagViewModel: VotedTagViewModel,
-        storeDetailInfoViewModel: StoreDetailInfoViewModel,
-        productListViewModel: ProductListViewModel,
-        operationInfoViewModel: OperationInfoViewModel
-    ) {
-        self.detailReviewViewModel = detailReviewViewModel
-        self.votedTagViewModel = votedTagViewModel
-        self.storeDetailInfoViewModel = storeDetailInfoViewModel
-        self.productListViewModel = productListViewModel
-        self.operationInfoViewModel = operationInfoViewModel
+    private func cancelFetchingProductList() {
+        productListLoadTask?.cancel()
     }
 }
 
+// MARK: - Enums
 extension StoreDetailViewModel {
-    enum Mode {
+    enum TabBarMode {
         case productLists
         case reviews
         case operationInfo
@@ -92,127 +84,31 @@ extension StoreDetailViewModel {
         }
     }
 
-    enum StoreInfoSection {
-        case main
+    enum StoreInfoButtonType {
+        case phone
+        case link
+        case like
 
-        var section: NSCollectionLayoutSection {
-            let item = NSCollectionLayoutItem(
-                layoutSize: .init(widthDimension: .fractionalWidth(1),
-                                  heightDimension: .estimated(400))
-            )
-            let group = NSCollectionLayoutGroup.vertical(
-                layoutSize: .init(widthDimension: .fractionalWidth(1),
-                                  heightDimension: .estimated(400)), subitems: [item]
-            )
-            let section = NSCollectionLayoutSection(group: group)
-            return section
-        }
-    }
-
-    enum ProductListSection: Int, CaseIterable {
-        case productCategory
-        case productsCount
-        case productList
-
-        var cellHeight: CGFloat {
-            return 113
-        }
-
-        var reuseIdentifier: String {
+        var image: UIImage? {
             switch self {
-            case .productCategory:
-                return ProductCategoriesCell.reuseIdentifier
-            case .productsCount:
-                return ProductListHeaderCell.reuseIdentifier
-            case .productList:
-                return ProductCell.reuseIdentifier
+            case .phone:
+                return UIImage(systemName: "phone")
+            case .link:
+                return UIImage(systemName: "link")
+            case .like:
+                return UIImage(systemName: "hand.thumbsup")
             }
         }
 
-        var cell: UICollectionViewCell.Type {
+        var title: String {
             switch self {
-            case .productCategory:
-                return ProductCategoriesCell.self
-            case .productsCount:
-                return ProductListHeaderCell.self
-            case .productList:
-                return ProductCell.self
+            case .phone:
+                return "전화"
+            case .link:
+                return "매장"
+            case .like:
+                return "추천"
             }
-        }
-
-        var section: NSCollectionLayoutSection {
-            let item = NSCollectionLayoutItem(layoutSize: .init(widthDimension: .fractionalWidth(1), heightDimension: .estimated(cellHeight)))
-            let group = NSCollectionLayoutGroup.vertical(layoutSize: .init(widthDimension: .fractionalWidth(1), heightDimension: .estimated(cellHeight)), subitems: [item])
-            let section = NSCollectionLayoutSection(group: group)
-            section.contentInsets = .init(top: 0, leading: 16, bottom: 0, trailing: 16)
-            return section
-        }
-    }
-
-    enum ReviewSection: Int, CaseIterable {
-        case moveToWriteReview = 1
-        case firstReviewRequest = 2
-        case votedCount = 3
-        case votedTag = 4
-        case detailReviewCount = 5
-        case detailReviews = 6
-
-        var cellHeight: CGFloat {
-            return 600
-        }
-
-        var reuseIdentifier: String {
-            switch self {
-            case .moveToWriteReview:
-                return MoveToWriteReviewCell.reuseIdentifier
-            case .firstReviewRequest:
-                return FirstReviewRequestCell.reuseIdentifier
-            case .votedCount:
-                return VotedCountLabelCell.reuseIdentifier
-            case .votedTag:
-                return VotedTagCell.reuseIdentifier
-            case .detailReviewCount:
-                return DetailReviewCountCell.reuseIdentifier
-            case .detailReviews:
-                return DetailReviewCell.reuseIdentifier
-            }
-        }
-
-        var cell: UICollectionViewCell.Type {
-            switch self {
-            case .moveToWriteReview:
-                return MoveToWriteReviewCell.self
-            case .firstReviewRequest:
-                return FirstReviewRequestCell.self
-            case .votedCount:
-                return VotedCountLabelCell.self
-            case .votedTag:
-                return VotedTagCell.self
-            case .detailReviewCount:
-                return DetailReviewCountCell.self
-            case .detailReviews:
-                return DetailReviewCell.self
-            }
-        }
-
-        var section: NSCollectionLayoutSection {
-            let item = NSCollectionLayoutItem(layoutSize: .init(widthDimension: .fractionalWidth(1), heightDimension: .estimated(cellHeight)))
-            let group = NSCollectionLayoutGroup.vertical(layoutSize: .init(widthDimension: .fractionalWidth(1), heightDimension: .estimated(cellHeight)), subitems: [item])
-            let section = NSCollectionLayoutSection(group: group)
-            section.contentInsets = .init(top: 0, leading: 16, bottom: 0, trailing: 16)
-            return section
-        }
-    }
-
-    enum OperationInfoSection {
-        case main
-
-        var section: NSCollectionLayoutSection {
-            let item = NSCollectionLayoutItem(layoutSize: .init(widthDimension: .fractionalWidth(1), heightDimension: .estimated(500)))
-            let group = NSCollectionLayoutGroup.vertical(layoutSize: .init(widthDimension: .fractionalWidth(1), heightDimension: .estimated(500)), subitems: [item])
-            let section = NSCollectionLayoutSection(group: group)
-            section.contentInsets = .init(top: 0, leading: 16, bottom: 0, trailing: 16)
-            return section
         }
     }
 }
