--- RefillStation/RefillStation/Presentation/StoreDetailScene/ViewModel/VotedTagViewModel.swift
@@ -1,14 +1,0 @@
-//
-//  VotedCollectionViewModel.swift
-//  RefillStation
-//
-//  Created by 천수현 on 2022/11/23.
-//
-
-import Foundation
-
-final class VotedTagViewModel {
-
-    var totalVoteCount = 5
-    var tagReviews = MockEntityData.tagReviews()
-}
